+++ internal/impl/sql/input_sql_raw_test.go
@@ -24,7 +24,7 @@ args_mapping: 'root = [ this.id ]'
 	selectConfig, err := spec.ParseYAML(conf, env)
 	require.NoError(t, err)
 
-	selectInput, err := newSQLRawInputFromConfig(selectConfig, nil)
+	selectInput, err := newSQLRawInputFromConfig(selectConfig, service.MockResources())
 	require.NoError(t, err)
 	require.NoError(t, selectInput.Close(context.Background()))
 }