+++ internal/impl/sql/output_sql_deprecated.go
@@ -40,7 +40,7 @@ func init() {
 			if maxInFlight, err = conf.FieldInt("max_in_flight"); err != nil {
 				return
 			}
-			out, err = newSQLDeprecatedOutputFromConfig(conf, mgr.Logger())
+			out, err = newSQLDeprecatedOutputFromConfig(conf, mgr)
 			return
 		})
 	if err != nil {
@@ -50,7 +50,7 @@ func init() {
 
 //------------------------------------------------------------------------------
 
-func newSQLDeprecatedOutputFromConfig(conf *service.ParsedConfig, logger *service.Logger) (*sqlRawOutput, error) {
+func newSQLDeprecatedOutputFromConfig(conf *service.ParsedConfig, mgr *service.Resources) (*sqlRawOutput, error) {
 	driverStr, err := conf.FieldString("driver")
 	if err != nil {
 		return nil, err
@@ -73,9 +73,9 @@ func newSQLDeprecatedOutputFromConfig(conf *service.ParsedConfig, logger *servic
 		}
 	}
 
-	connSettings, err := connSettingsFromParsed(conf)
+	connSettings, err := connSettingsFromParsed(conf, mgr)
 	if err != nil {
 		return nil, err
 	}
-	return newSQLRawOutput(logger, driverStr, dsnStr, queryStatic, nil, argsMapping, connSettings), nil
+	return newSQLRawOutput(mgr.Logger(), driverStr, dsnStr, queryStatic, nil, argsMapping, connSettings), nil
 }