+++ CHANGELOG.md
@@ -19,6 +19,7 @@ All notable changes to this project will be documented in this file.
 - New `cassandra` input.
 - New `base64_encode` and `base64_decode` functions for the awk processor.
 - Param `use_number` added to the `parse_json` bloblang method.
+- Fields `init_statement` and `init_files` added to all sql components.
 
 ### Fixed